Homework 3
I converged on a solution for around half of the problems with newton'ss
method and by changing the negative singular values of the hessian to 0,
the algo ran for all but two or so of the problems but did not converge
on a solution, and I got errors on about two of the problems and they
would not run, I am not sure why.

I could not get the newton method to work on the logistic regression. .
. something with the hessian? ? ? the gradient went to NaN after
itteration 1.